+++ src/Rendering/Uniform.hpp
@@ -97,11 +97,11 @@ struct TextureUniform : ShaderUniform
 
 struct UniformList
 {
-	unsigned int uniformDataSize; // CPU side
-	unsigned int uniformBufferSize; // GPU side
+	size_t uniformDataSize; // CPU side
+	size_t uniformBufferSize; // GPU side
 
-	unsigned int bufferUniformCount;
-	unsigned int textureUniformCount;
+	size_t bufferUniformCount;
+	size_t textureUniformCount;
 	BufferUniform* bufferUniforms;
 	TextureUniform* textureUniforms;
